Refactor decorators to be included as modules, rather than subclasses of SimpleDelegator.
Part of the work to optimise the loading time to pages.

It wasn't possible to include all the bookings of a trip when we had to then instantiate
decorators for each booking.

This change means that the "decorators" are just included as modules, mixed in the
underlying model (bookings and trips).

Follow up work will optimise the page load time using an includes in the ActiveRecord/
SQL statement.