Add has_minimum_number_of_guests? to trip model.
Part of the work to allow the payment of deposits and then the subsequent automatic payment of the remainder at a certain date.

has_minimum_number_of_guests? is used to check to take full payment from a trip's bookings
or to send the refund all bookings email to the guide (in subsequent work).

This is one part that can be broken out of the main work into digestible pull requests.

More to follow...