Fix bug in paying_deposit? method in Bookings::Payment serive model.
Part of the work to allow the payment of deposits and then the subsequent
automatic payment of the remainder at a certain date.

When working on that task, the #paying_deposit? method was returning true
when we were taking the remainder of the trip cost (ie: full_cost - deposit).
So this was breaking the logic further up the stack.

This fix means that #paying_deposit? will only return true if the payment
now due, is EXACTLY the same as the deposit cost of the trip.

This change is covered with subsequent integration tests, in follow up work.

Also after looking in Stripe dashboard, it seemed better to add '(deposit)'
to the description of the charge when it was a deposit payment.

This is one part that can be broken out of the main work into digestible pull requests.

More to follow...