Allow Bookings::CostCalculator to handle remaining cost due after deposit is paid.
Part of the work to allow the payment of deposits and then the subsequent automatic
payment of the remainder at a certain date.

We need to calculate the amount_due for bookings that have paid a deposit and are now due
to pay the remaining cost.

This required making #total_paid method public in the Bookings::PaymentStatus service model.

This is one part that can be broken out of the main work into digestible pull requests.

More to follow...