Add PayOutstandingTripCostJob and task.
(Almost) final part of the work to automatically take the oustanding
cost of the trip from bookings that have only paid the deposit...
OR, if not enough guests on the trip, send the Guide an email to alert
them and confirm they want to refund the deposits to all guests.

Follow up work will actually add the email that the guide should receive
when there's not enough guests on trip.