Refactor trip factory and related specs.
Part of the work to send an email to a Guide when we want to cancel
a trip. That work showed the current test set up was broken.

The should always be an organisation_membership created for the guide
of who is the owner of the trip.

This moves this logic into the factory.