Send support email when trip needs to be cancelled.
Part of the work to automatically take outstanding payment or
refund guests' deposits when trip deadline has approached.

We already send the trip's guide an email telling them they need
to cancel the trip and refund deposits, but if they don't carry
out this action then we need to be aware and chase it up.

Giving the guide three days to refund the deposits means if they
get the cancel trip email over the weekend they still have time to
act before we chase them.