Add refunded status to bookings.
This is to differentiate between bookings that are live and those
that have been refunded, after a trip has been cancelled, etc.

Follow up work will use Stripe Webhooks to automatically update this
field when there's been a refund.

Also, will need something in the UI to show this booking has been refunded
for the guide to see.