Refactor public bookings forms.
This initially was to fix the routing issue when there was a problem with
creating a new booking (which meant the form could never be sent again when
a new booking corrected any issues with their email, card no, etc) BUT
once that was fixed it became apparent that these forms could be simplified.

We don't need to pass in the url as a param anymore, for ex.

The original fix required adding method: :post to the new form.
Without this, as the booking was now present, on error, it was trying to post
a patch/update - when it should be a post/ create call - when the new booking
amended any errors in the form.