Fix seeds.rb.
Since allergies and dietary conditions have become has_many associations to bookings,
the seeds file was broken as it was still treating them as enumns on the booking model.

Also, now we process the trip full price, multiplying it by 100 to turn it into a Stripe
friendly cents version, the prices of the trips' full_cost was 100x too large.

Also, moved the guide's password into an env var. - More proffessional.