Fix email regex.
When I was making a new booking I was blocked from using an email address
with a + in and an email address with a domain of sussex.ac.uk.

We want to allow these types of email addresses, so I have made the regex
matching valid emails more liberal to include these types.