Remove validations that now fail when a new organisation is created.
Part of the work to automatically create an organisation when a guide
signs up.

As we don't have a way (yet) to allow new guides to set the currency and name
of their new organisation on creation, we need to remove these validations on
those fields so that new organisations can actually be created.