Use a default currency (EUR) when crating trips when their organisation has no currency set.
Part of the work to automatically create organisations when a guide signs up.

An app error occured when a guide went to create a new trip but they hadn't (as it's not
currently possible to) add a currency to their organisation.

If a trip doesn't have currency (ie: as when you are creating a trip) then we would provide
the trip's organisation's currency as a default.

But now, we create an organisation, we do not add a default currency to it, so this was
causing a nil error when we tried to reference it.

Further work can determine the guide's country and currency when they are created and assign
this to their organisation.