Fix booking's other_information bug.
New bookings were showing up as having other_information when they hadn't
entered anything for this field.

The field was being passed as a blank string which set the field to a
blank string and made it "truthy" in the views where it was checked
for haivng a presence.

Now the controller strips any params that are blank.

ref: https://github.com/AlanDonohoe/mybookingsherpa/issues/409