An organisation has_one subscription now, not many.
Part of the work to onboard new guides.

The check to see if an organisation is on_trial or not is a bit odd.
It's using a check on a possible, as yet, unused, associations, the
subscription, to decide if the organisation is on trial or not.

We can make this better. But for now, let's make sure that an org
can only have one subscription. This makes sense, and makes the test
suite pass...