Show you're in trial banner if organisation created over 1 day ago.
Part of the work to onboard new guides.

If a guide NEVER creates a test booking, they will be blocked from
proceeding.

This allows them a way out. After 24 hours of them signing up, the
in trial banner will appear irregardless of what they've done (as long
as they're still in trial).