Add complete field to onboarding model.
Part of the work to onboard new guides.

Based on certain actions taken by the new guide, their organisation
can be marked as having their onboarding (process) complete. Then we
don't show all the onboarding messages to them and their use the live
Stripe API to make real payments.

Because Rails is lovely, we get to add a boolean field to an organisation's
onboarding model, that adds the helper method 'complete?' to onboarding
model, then we can delegate the call to 'complete?' in the organisation
model to the onboarding model, WITH the model's prefix, so then we can just
call this on an organisation to determine if it's completed it's onboarding
process:

organisation.onboarding_complete?