Make sure the correct Stripe API key is being used (conditionally) in the backend of app.
Part of the work to onboard new guides.

We need to send Stripe the live key when we are REALLY onboarding new guides
and send Stripe the test key when we're in test mode.

Further work is required to make sure we're using the correct key in the front end
but that will require turning the existing Stripe JS into a Stimulus controller.