Only send Stripe account token to our servers, not all of company details.
Part of the work to onboard new guides who are companies (not individuals).

This work creates a hidden form that has a single hidden field, which gets
set to the value of the Stripe account id whihc Stripe's JS API returns.
This form is then sent to our server and the account id is used to make a
Stripe account via Stripe's ruby API. Keeping all sensitive company data
away from our servers.