Send company tax/ vat id to Stripe JS when creating new Stripe Account.
Part of the work to onboard new guides who are companies.

May need to tweak this as according to the Stripe API docs*, the tax
id may be the company number (in the UK) and other numbers in other
countries.

So a company in the UK could have both a tax number (= company number)
and a VAT number....

So needs more work in this area... Also need to check in Stripe dashboard
when these accounts are created, what's the bare minimum required data and
only ask for that from new guides.