Refactor hide_related_element_controller -> onboarding_controller
This work started with a change to the html to hide the final onboarding
element when it was clicked. But I couldn't find the JS controller that
handled the this... which led me to think it wasn't named very well (by
myself, previously).

So I thought I'd rename it to be the onboarding controller and we can add
all onboarding specifc JS into this controller.