Use separate fields to persist Stripe account_id when it's the live and test account ids.
Part of the work to use Stripe's hosted onboarding for new guides.

We will need to have access to the live Stripe account id to make the account link
which new guides will use to onboard.

If we don't do this, then we will need to make a new call to Stripe API every time
the welcome view is loaded to create a new account, which is very wasteful and time
consuming in a web request.