Remove Stripe client_secret from any logging.
Part of the work to start using the new Stripe SCA API.

We want to moke sure we're not sharing this sensitive info anywhere.

ref: https://stripe.com/docs/payments/payment-intents/web#passing-to-client

The client secret can be used to complete the payment process with the amount specified on the PaymentIntent. It should not be logged, embedded in URLs, or exposed to anyone other than the customer. Make sure that you have TLS enabled on any page that includes the client secret