Make payments' booking association optional.
Part of the work to use Stripe's new SCA API.

Now payments may be created before their associated booking is created,
this is in the Stripe payment_intents webhook.

It's that quick! We submit the payment_intent to be created from the client
to Stripe's API, then it sends a webhook back with the payment_intent status.

It also posts back a payment_intent id to our form in the client, which is
then posted to our bookings create controller action. It's here that the
booking is created but if this request is AFTER the web request, then we
need to create the payment first (in the webhook) and then associate it
with the new booking AFTER.

That work is to follow (ie: the updated bookings#create controller/action).