Relax trip validations.
We can allow guides to create trips with all sorts of characters in...

May need to check that the slug generated from the name can handle odd characters.

Also allow the start date to be on the same day as the end date - for trips that only
run for 1 day.