Stress the optional text for the deposit field/label in new trip form.
Part of the work post first user testing of new guide onboarding.

The fact that it was optional was missed by one of our user testers.

Further work could improve this quick fix/hack by moving the deposit
field into a an accordian? which you'd have to click open to access...

But this will hopefully do for now.