Make call to refresh_stripe_account_capabilities when viewing new bank account view.
Part of the work post first user testing of new guides onboarding.

There was a race condition that occured when the webhook letting us know the
Stripe account onboarding had all been completed came in AFTER we were redirected
back to the app.

This caused the new bank page to redirect to the trips index, when it should of
remained on this page as the new guide needs to add their bank details.

This fixes that by making one final check that the stripe account has had its
charges enabled before determining whether to redirect or not.