Refactor copy and UX of in trial new booking modals.
Part of the work post first user testing of new guide onboarding process.

I think this is an improvement.

The main difference is the final modal appears after a 5 second delay,
which allows the user to see the final page that the guest would see, then
follow this with the "All done" modal and a link back to "My trips".

This feels loads better.