Remove explainer pointers and text from in trial example booking walkthrough.
Part of the work post first user testing of new guide onboarding.

There was confusion about what the guest would see and what the guide would see.

This way, there's nothing on the page that is for only meant for the guide to see.

All the guide related content and explainer elements are in modals.

A lot clearer.