Remove all of in trial badge.
Missed some of this in a previous commit.

Part of the onboarding work. We will leave this badge off for now
as it violates the other aim we were trying to acheive - remove ALL
guide related content from the guest booking views - this came up
as something that was confusing in user testing.