Show our and Stripe's fees to guides at end of onboarding process.
We don't want to implement any "dark patterns" and hide away the costs of the service
in some small print.

But guides don't need to know/ be reminded of this stuff really... until they've completed
onboarding.

There's two sets of pricing for Stripe:

1) Regular: https://stripe.com/gb/pricing
2) Connect: https://stripe.com/gb/connect/pricing

I am assuming that the guides get charged using the (1) Regular fees
and we, as a connect platform, get charged as well on the (2) Connect fees.

Will need to test this in live mode...