Add off session failed payment email (and required routes, etc).
Part of the work to bring our payment system in line with Stripe's
SCA requirements.

We need to alert (via these emails) guests and guides, when the attempt
to collect the outstanding amount has failed.

The guest can then click the link in the email and make full payment for
the failed amount.

Follow up work will add the views and controllers which will handle the
UX of allowing guests to update their card details and settle outstanding
amounts.