Hard code host in links of final page of onboarding.
Final tweak to fix the odd subdomain redirect bugs in production.

We needed to explicitly state that we wanted to link to the host: BASE_DOMAIN, which
will be localhost:5000, or app.mybookingsherpa.com, depending on the environment.

Otherwise we will be linking to www.bookyour.place/trips/index - which would mean the guide
would be logged out on that domain...