Fix public booking link on staging.
On all environments other than staging (ie: local and production),
we can use an organisation's subdomain or www (if they don't have a
subdomain).

This doesn't work on staging as we don't have the dns set up for this,
to handle www / subdomain.

This fixes that by not prefixing the www or sudomain to the public
booking link when we know we are on staging.