Temporarily turn off rails db:migrate in heroku release-tasks.sh
Attempt to fix a migration error.

When onboardings table was created there wasn't a ref to organisations
but subsequently this was created somehow but without having a migration
to handle it.

The last migation added this reference but when rails db:migrate is run
this error is thrown:

PG::DuplicateColumn: ERROR:  column "organisation_id" of relation "onboardings" already exists

This will hopefully fix this by just loading the schema rather than running the migrations.

Will need to run: heroku run rake db:schema:load -a app_name

And then will need to revert to rails db:migrate afterwards in release-tasks.sh