Fix flakey tests
These tests fail when the randomly generated names for trips and guests
include an apostrophe, such as Helen O'Mara, etc.

This is because it is encoded in the page of the html to &#39;

Then the test is looking for the presence of Helen O'Mara, when it
is present as Helen O&#39;Mara.

This escapes the guest and trip names so that it is looking for the
html-encoded version of the names.