Attempt to fix broken in trial public booking links.
Part of the work post using real data on staging / pre testing with Arno.

Because staging doesn't have DNS set up to handle subdomains we treat it a bit
diferrently to local dev and production.

But this meant the URLs for the public booking edit and show views were breaking.

This should hopefully fix this.