Send failed payment emails.
Part of the work to deal with SCA when an off-session payment, ie:
for the remaining amount of the trip, when only a deposit has been
paid, fails.

We need to notify the guest and the guide and send them both emails
with a link to update the guest's card details for the outstanding amount.