Make the destination fee 1% of the full cost of the trip.
There was a bug when we ran the pay_outstanding_trip_cost_job, which
calculated the destination fee as 1% of the outstanding amount, not
the full_cost.

This now calculates the destination fee 1% of the full cost of the trip
both in the pay_outstanding_trip_cost_job and the Bookings::PaymentIntents
(full payment upfront) model.