Fix bug where payment method was always being retreived using test api key.
Part of the work of testing production with real, live payments.

The pay_outstanding_trip_cost_job was failing because it was trying to
retreive a payment_method in the live API but always using the test API key.

This was due to not passing in the use_test_api: param and then the receiver
then using the default value for this keyword param = true.