Remove guest sign in / sign up routes.
Part of the work to get ready for Launch Jan 2020.

Having these routes open is a security risk.

We need to make sure that if someone who makes a booking with an email address,
then a hacker can't sign up for an guest account and then access that booking's
data.