Round up deposit amount due to nearest whole number.
Part of the work to tidy up how we handle deposits.

This means the deposit and outstanding amount (and trip full cost)
will only ever have cents values of 00.

This will tidy up the UI in this area.

I moved the logic for calculating the deposit_cost from the Trip model
into a specialised DepositCalculator service model.