Fix booking payment status showing deposit_paid when it should be fully paid, bug.
This was a race condition where the stripe payment_intent webhook returned with a
successful payment_intent status before the booking has finished being created in
the Public::Trips::BookingsController.

This caused a payment to be made (in the webhook) that could not be associated with
a booking. So there was a stray payment created, which meant the booking only ever
had the payment created in the Public::Trips::BookingsController without its status
ever being updated, hence this bug.

Work had to extend out to accomodate the changes to make sure the
PayOutstandingTripCostJob worked too.

Further work will reinstate the pending tests to make sure there's no regression bugs
been created.

ref:
https://3.basecamp.com/4363532/buckets/14684619/messages/2263335533#__recording_2353252038