Fix pay outstanding amount job when processed with failing/ auth required cards.
This was a bug that occured with a recent refactor in the webhook controller
that handled Stripe's PaymentIntents' updates.

When the Bookings::PayOutstandingTripCostJob was run and failed to create
a payment record associated with the booking, because the creation of the
new payment_intent failed due to a card error/ authorisation required error,
then in the webhook post alerting us that the payment_intent status was
failed, there was a payment_intent id, but it wasn't associated with a payment
record on our side (as that creation blew up with the Stripe card error).

So we need to find_or_create_by(stripe_payment_intent_id: stripe_payment_intent_id)
a payment and associate it with the booking, then send out the failed payment emails.

There is also some refactoring of files that are in this area:
- Give a default value for the payment_failure_message param, instead of a conditional in the controller

- Deletion of tests for cases that don't actually happen in this app:
  - Failed payment_intent without a booking (as we only can get a failed payment_intent
  in the Bookings::PayOutstandingTripCostJob, which will ALWAYS have a booking)
  - With a failed in_session SCA attempt: a failed SCA attempt will only happen in the
  Bookings::PayOutstandingTripCostJob, which is always off_session.