Stop checking that Payment with payment_intent_id exists in Stripe payment_intents webhook controller.
This was added to stop errors when Stripe posted to development and staging webhooks,
when it was trying to update a payment on staging, that actually existed in development.

BUT... having this check only worked on successful payments. When there was an attempt to
make a payment that failed (in the update_successful_payment_job), then the payment was
never created.

Then when Stripe posted back to this webhook, with payment_intent status = failed, this
check failed and returned a not_found status code. When really it should create a payment
with a failed status.