Find_or_create payment by stripe_payment_intent_id, not just find.
When there's a failed attempt at making a payment (in the PayOutstandingTripCostJob)
then a payment is not created at all... as the Stripe::CardError is thrown and this
makes the job explode... resulting in no payment being created.

When that happens, and a guest has the option to update their card payment details,
the webhook fired back to us from Stripe is actually a payment_intent status = success.

So now we need to CREATE the payment, as it doens't exist already.

It would of only existed when it was a regular successful payment process (in session).

We could refactor this so that we handle the failed payment attempt more gracefully and actually
do create the payment with a failed status.