Check if post to webhook is actually meant for dev or staging.
Getting all sorts of app errors which are red herrings.

Passing in the base_domain to the stripe API call in the
payment_intent's creation metadata, we can then check in the
resultant webhook post via the not_meant_for_this_environment?
method that the payment_intent came from this environment or another
(dev or staging).

And then only continue to process the payment_intent from there.

This will quiet down RollBar and only raise genuine errors from
now on.