Bug fix: Prevent duplicate new booking payments.
The new public booking form was being submitted to Stripe.js API
without checking the validity of the rest of the form (it only checks
the format of the card element).

When that was happening and the booking creation failed in our controller
the user was redirected back to the new booking form, asked to correct the
invalid form fields and then got to submit the form again... which was
then hitting Stripe for a new booking.

This work uses client side validation, using the same regex as the server
side checks for valid name and email address. So prevents the form being
submitted to Stripe before these are all valid.