Replace custom.js with form_validation_controller.js
Moves the Bootstrap form validation JS into a Stimulus controller.

This follows on from work in this area that was allowing the new booking
form to be submitted to Stripe even when there was validation errors in
some of the form fields.

Further work will need to add this to all forms that require client side
validation.