Fix guides bank account form bug.
The work removing some of the form validation from certain forms
that already had their own validation within their specific JS
controllers also accidentally added back the stripe--bank-account
controller to the guides' bank account form.

The controller was already attached to this page higher up in an
enclosing div, so this casued the JS associated with this form to
break.

This fixes that.