Fix bug when a trip has a deposit of 50% which caused the pay outstanding job to fail.
The check to see if we were only_paying_deposit? was returning true, when it should of
been false, when a booking was for a trip that had a 50% deposit. This was because we
were the check in this method was only checking for to see if the amount_due equaled the
deposit_cost.

Which would be true, for a 50% deposit trip, even if the deposit has been paid.

So in the request to Stripe to make a payment intent we got this error:

"Stripe::InvalidRequestError: You cannot confirm with the `off_session` parameter when
`setup_future_usage` is also set on the PaymentIntent because you cannot set up future
usages when processing payment..."

As we we thought we were taking the deposit payment (so setup_future_usage = true) AND
we were taking the outstanding amount (off_session = true). Kaboom!

This should fix that.