Force trips with deposit to have deposit and full payment window.
Whilst working in this area, with a related bug around deposits, it
didn't make any sense for a trip to have a deposit but not a full
payment window.

It allowed for conflicting behaviour in the test suite, so made a call
to enforce the presence of both deposit and full payment window or none at all.