Set onboarding to complete in the Stripe webhook not the bank accoutn controller.
This is to deal with how Stripe Connect Hosted Onboarding now seem to
take a while to verify a new users' ID.

Previously we could just check, when the user returned from the hosted
onboarding, if they had completed everything or not.

But now the user can add all their details to the Stripe onboarding but
then Stripe takes a little while before finally requesting proof of ID
and address documents.

So now, we can only truely know when onboarding is complete when we
receive the webhook from Stripe with this information in it.